Split runner registration tests
